Because of addition of the Rebin action, listener is modified to take floating point numbers instead of just integers.

+++ jam/global/CommandListener.java
@@ -15,5 +15,5 @@ package jam.global;
 	 * @param command word indecating action to take
 	 * @param parameters list of typed numbers which are parameters to the command
 	 */
-	public void commandPerform(String command, int[] parameters);	
+	public void commandPerform(String command, double [] parameters);	
 }
